getChartInfo: add version check and test, move tests
