+++ internal/app/state_test.go
@@ -372,3 +372,127 @@ func Test_state_validate(t *testing.T) {
 		})
 	}
 }
+
+func createFullReleasePointer(chart, version string) *release {
+	return &release{
+		Name:         "",
+		Description:  "",
+		Namespace:    "",
+		Enabled:      true,
+		Chart:        chart,
+		Version:      version,
+		ValuesFile:   "",
+		ValuesFiles:  []string{},
+		SecretsFile:  "",
+		SecretsFiles: []string{},
+		Test:         false,
+		Protected:    false,
+		Wait:         false,
+		Priority:     0,
+		Set:          make(map[string]string),
+		SetString:    make(map[string]string),
+		HelmFlags:    []string{},
+		NoHooks:      false,
+		Timeout:      0,
+		PostRenderer: "",
+	}
+}
+
+func Test_state_getReleaseChartsInfo(t *testing.T) {
+	type args struct {
+		apps map[string]*release
+	}
+
+	tests := []struct {
+		name       string
+		targetFlag []string
+		groupFlag  []string
+		args       args
+		want       bool
+	}{
+		{
+			name: "test case 1: valid local path with no chart",
+			args: args{
+				apps: map[string]*release{
+					"app": createFullReleasePointer(os.TempDir()+"/helmsman-tests/myapp", ""),
+				},
+			},
+			want: false,
+		}, {
+			name: "test case 2: invalid local path",
+			args: args{
+				apps: map[string]*release{
+					"app": createFullReleasePointer(os.TempDir()+"/does-not-exist/myapp", ""),
+				},
+			},
+			want: false,
+		}, {
+			name: "test case 3: valid chart local path with whitespace",
+			args: args{
+				apps: map[string]*release{
+					"app": createFullReleasePointer(os.TempDir()+"/helmsman-tests/dir-with space/myapp", "0.1.0"),
+				},
+			},
+			want: true,
+		}, {
+			name: "test case 4: valid chart from repo",
+			args: args{
+				apps: map[string]*release{
+					"app": createFullReleasePointer("prometheus-community/prometheus", "11.16.5"),
+				},
+			},
+			want: true,
+		}, {
+			name:       "test case 5: invalid local path for chart ignored with -target flag, while other app was targeted",
+			targetFlag: []string{"notThisOne"},
+			args: args{
+				apps: map[string]*release{
+					"app": createFullReleasePointer(os.TempDir()+"/does-not-exist/myapp", ""),
+				},
+			},
+			want: true,
+		}, {
+			name:       "test case 6: invalid local path for chart included with -target flag",
+			targetFlag: []string{"app"},
+			args: args{
+				apps: map[string]*release{
+					"app": createFullReleasePointer(os.TempDir()+"/does-not-exist/myapp", ""),
+				},
+			},
+			want: false,
+		}, {
+			name:       "test case 7: multiple valid local apps with the same chart version",
+			targetFlag: []string{"app"},
+			args: args{
+				apps: map[string]*release{
+					"app1": createFullReleasePointer(os.TempDir()+"/helmsman-tests/dir-with space/myapp", "0.1.0"),
+					"app2": createFullReleasePointer(os.TempDir()+"/helmsman-tests/dir-with space/myapp", "0.1.0"),
+					"app3": createFullReleasePointer(os.TempDir()+"/helmsman-tests/dir-with space/myapp", "0.1.0"),
+					"app4": createFullReleasePointer(os.TempDir()+"/helmsman-tests/dir-with space/myapp", "0.1.0"),
+					"app5": createFullReleasePointer(os.TempDir()+"/helmsman-tests/dir-with space/myapp", "0.1.0"),
+				},
+			},
+			want: true,
+		},
+	}
+
+	teardownTestCase := setupTestCase(t)
+	defer teardownTestCase(t)
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			stt := &state{Apps: tt.args.apps}
+			stt.disableUntargetedApps(tt.groupFlag, tt.targetFlag)
+			err := stt.getReleaseChartsInfo()
+			switch err.(type) {
+			case nil:
+				if tt.want != true {
+					t.Errorf("getReleaseChartsInfo() = %v, want error", err)
+				}
+			case error:
+				if tt.want != false {
+					t.Errorf("getReleaseChartsInfo() = %v, want nil", err)
+				}
+			}
+		})
+	}
+}
